Fix filter never bypassed
isBypassed only deals with whether the device is on/off. When the device
is on but the cutoff is set to 0, we also need to bypass the signal 
processing. 

Previously, when cutoff is 0, it will run hi-pass with cutoff = 20 (which
will probably just cause phase change and not really make any audible 
frequency change)